Bound the remaining reads sized by untrusted counts
Reviewing the load command bounds turned up three more of the same kind elsewhere.

A symbol, indirect symbol or relocation count was multiplied by an entry size in 32-bit arithmetic. A count large enough to wrap that product gave a small, plausible length that passed the coverage check, and the loop then read past what had been read. Those products are now computed in 64 bits, so the wrap cannot happen and an impossible length is reported instead. The symbol table count reached the caller as an ArgumentOutOfRangeException rather than a diagnostic.

A universal binary header's slice count was used to size an allocation before anything checked the file was big enough to hold that many. It is now checked against the bytes actually there.

Signing changed a good deal before it could fail: it removed the previous signature's content, moved the command, resized __LINKEDIT and cleared the stale flag, then wrote the image to take the digests, and that write can fail. What it changes is now put back if it does, so a failed signing leaves the image exactly as it was rather than half signed.

- src/LibObjectFile/MachO/MachOFile.Symbols.cs: lengths computed in 64 bits
- src/LibObjectFile/MachO/MachOFatFile.cs: the slice count checked before it sizes anything
- src/LibObjectFile/MachO/MachOFile.Sign.cs: signing restores what it changed if it fails

+++ src/LibObjectFile/MachO/MachOFile.Relocations.cs
@@ -30,7 +30,7 @@ partial class MachOFile
 
         if (section.NumberOfRelocations == 0) return [];
 
-        var bytes = ReadFileBytes(section.RelocationOffset, section.NumberOfRelocations * MachORelocation.EntrySize, $"relocations of {section.SegmentName},{section.Name}");
+        var bytes = ReadFileBytes(section.RelocationOffset, (ulong)section.NumberOfRelocations * MachORelocation.EntrySize, $"relocations of {section.SegmentName},{section.Name}");
         var relocations = new MachORelocation[section.NumberOfRelocations];
         for (var i = 0; i < relocations.Length; i++)
         {
